+++ app/wyzecam/api.py
@@ -6,8 +6,8 @@ from typing import Any, Dict, List, Optional
 import requests
 from wyzecam.api_models import WyzeAccount, WyzeCamera, WyzeCredential
 
-IOS_VERSION = "16.1"
-APP_VERSION = "2.36.0.14"
+IOS_VERSION = "16.1.1"
+APP_VERSION = "2.37.0.12"
 
 SV_VALUE = "e1fe392906d54888a9b99b88de4162d7"
 SC_VALUE = "9f275790cab94a72bd206c8876429f3c"